Extract planner out to separate module
